Styling for navbar, and moved like button to the details page
