+++ src/components/Navigation/Navigation.css
@@ -1,3 +1,6 @@
-.navbar {
-    background-color: blue;
+Navbar {
+    color: white
+}
+.parkadvisor {
+    font-weight: bold;
 }
