+++ src/components/Navigation/index.js
@@ -15,12 +15,18 @@ export default function Navigation() {
   const loginLogoutControls = token ? <LoggedIn /> : <LoggedOut />;
 
   return (
-    <Navbar className="navbar" bg="light" expand="lg">
+    <Navbar 
+      className="navbar" 
+      bg="light" 
+      expand="lg" 
+      bg="success"
+      >
       <Navbar.Brand as={NavLink} to="/">
-        ParkAdvisor
+      <img src="https://img.icons8.com/ios-filled/50/000000/national-park.png"/> 
+      <span className="parkadvisor">ParkAdvisor</span>
       </Navbar.Brand>
-      <Navbar.Toggle aria-controls="basic-navbar-nav" />
-      <Navbar.Collapse id="basic-navbar-nav">
+      <Navbar.Toggle aria-controls="responsive-navbar-nav" />
+      <Navbar.Collapse id="responsive-navbar-nav">
         <Nav style={{ width: "100%" }} fill>
           <NavbarItem path="/" linkText="Home" />
           {token ? <NavbarItem path="/createPark" linkText="Add a park" /> : ""}
